ML/requirements.txt
Flask==3.0.3
Pillow==10.0.1
Pillow==10.3.0
ultralytics==8.2.36
